requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
scikit-learn
pydantic
